refactor JSONStandardParser.kt into /util
